back-end/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
dataclasses-json==0.5.14
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-appconf==1.0.6
django-cacheops==7.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-import-export==3.3.3
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-staticfiles==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
facebook-sdk==3.1.0
funcy==2.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.20.1
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-facebook-api==0.18.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
simplejson==3.19.2
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tablib==3.5.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
